+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Change "<- Back to App" button on the updater page to redirect to the settings page rather than the homepage
   * Update links on "About" settings tab to open in a new tab
   * Add scroll functionality to more modals, hide the scrollbars
+  * Handle errors better in the stream editing dialog
 * System
   * Fixed a bug where support tunnel addresses would be served when querying for amplipi.local
   * Stream validation for URLs has been made more robust